requirements.txt
Django==1.4.7
djangorestframework==2.4.3
rsa==3.1.4
django-cors-headers==0.13
requests==2.4.3
pycrypto==2.6.1
pytest
